+++ src/Udyam/components/globe/files/globe-data-min.json
